backend/lambda/resolution-worker/requirements.txt
blessed==1.20.0
boto3==1.35.34
botocore==1.35.34
certifi==2024.8.30
chalice==1.31.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.10
inquirer==2.10.1
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-editor==1.0.4
PyYAML==6.0.2
readchar==4.2.0
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
urllib3==1.26.20
wcwidth==0.2.13
